Move the Dart code to support the libraries layout
