+++ lib/audioplayers_api.dart
@@ -1,0 +1,4 @@
+export 'api/player_mode.dart';
+export 'api/player_state.dart';
+export 'api/playing_route.dart';
+export 'api/release_mode.dart';
