+++ lib/web/audioplayers_web.dart
@@ -1,7 +1,7 @@
 import 'dart:async';
 import 'dart:html';
 
-import 'package:audioplayers/release_mode.dart';
+import 'package:audioplayers/api/release_mode.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_web_plugins/flutter_web_plugins.dart';
